Add dual-license info to README.md
Summary:
From #3417 and after talking to both GitHub and our open source legal team, the recommended approach was to explicitly state the dual-license in the readme.

Changing the license files to accommodate the auto-detection is too much of a pain, would involve editing every code file header.
Closes https://github.com/facebook/rocksdb/pull/3541

Differential Revision: D7171111

Pulled By: gfosco

fbshipit-source-id: 0ee7b134446015228249efe991fa5e76526ca0b0